Refactored of flib_ipc from header only to compiled library. - Intermediate commit, will segfault in get_write_index_latest   when using -O3 and alignas(16) for DualIndex
