+++ app/flesnet/CMakeLists.txt
@@ -12,7 +12,7 @@ include_directories("${PROJECT_SOURCE_DIR}/lib/fles_core"
 
 add_executable(flesnet ${APP_SOURCES} ${APP_HEADERS})
 
-target_link_libraries(flesnet fles_core fles_rdma logging ibverbs rdmacm ${Boost_LIBRARIES})
+target_link_libraries(flesnet fles_core fles_rdma flib_ipc logging ibverbs rdmacm ${Boost_LIBRARIES})
 if(LIBNUMA AND USE_LIBNUMA)
     target_link_libraries(flesnet ${LIBNUMA})
 endif()
