+++ app/mstool/CMakeLists.txt
@@ -15,7 +15,7 @@ include_directories("${PROJECT_SOURCE_DIR}/lib/fles_ipc"
 
 add_executable(mstool ${APP_SOURCES} ${APP_HEADERS})
 
-target_link_libraries(mstool fles_ipc fles_core logging crcutil ${Boost_LIBRARIES} ${CMAKE_THREAD_LIBS_INIT})
+target_link_libraries(mstool fles_ipc fles_core flib_ipc logging crcutil ${Boost_LIBRARIES} ${CMAKE_THREAD_LIBS_INIT})
 target_link_libraries(mstool ${ZMQ_LIBRARIES})
 if(LIBNUMA AND USE_LIBNUMA)
     target_link_libraries(mstool ${LIBNUMA})
