+++ lib/flib_ipc/shm_channel_provider.hpp
@@ -6,11 +6,9 @@
 #include "DualRingBuffer.hpp"
 #include "log.hpp"
 #include "shm_channel.hpp"
-#include <boost/date_time/posix_time/posix_time_types.hpp>
+#include "shm_device.hpp"
 #include <boost/interprocess/managed_shared_memory.hpp>
-#include <boost/interprocess/sync/scoped_lock.hpp>
-#include <boost/lexical_cast.hpp>
-#include <string>
+#include <memory>
 
 namespace ip = boost::interprocess;
 
@@ -33,63 +31,20 @@ public:
                        shm_device* shm_dev,
                        size_t index,
                        size_t data_buffer_size_exp,
-                       size_t desc_buffer_size_exp)
-      : shm_dev_(shm_dev) {
-    // allocate buffers
-    void* data_buffer_raw =
-        shm_alloc(shm, data_buffer_size_exp, sizeof(T_DATA));
-    void* desc_buffer_raw =
-        shm_alloc(shm, desc_buffer_size_exp, sizeof(T_DESC));
+                       size_t desc_buffer_size_exp);
 
-    // constuct channel exchange object in shared memory
-    std::string channel_name =
-        "shm_channel_" + boost::lexical_cast<std::string>(index);
-    shm_ch_ = shm->construct<shm_channel>(channel_name.c_str())(
-        shm, data_buffer_raw, data_buffer_size_exp, sizeof(T_DATA),
-        desc_buffer_raw, desc_buffer_size_exp, sizeof(T_DESC));
-    set_write_index({0, 0});
+  DualIndex get_read_index() override;
 
-    // initialize buffer info
-    T_DATA* data_buffer = reinterpret_cast<T_DATA*>(data_buffer_raw);
-    T_DESC* desc_buffer = reinterpret_cast<T_DESC*>(desc_buffer_raw);
+  void set_write_index(DualIndex new_write_index) override;
 
-    data_buffer_view_ = std::unique_ptr<RingBufferView<T_DATA>>(
-        new RingBufferView<T_DATA>(data_buffer, data_buffer_size_exp));
-    desc_buffer_view_ = std::unique_ptr<RingBufferView<T_DESC>>(
-        new RingBufferView<T_DESC>(desc_buffer, desc_buffer_size_exp));
-  }
+  void set_eof(bool eof) override;
 
-  DualIndex get_read_index() override {
-    ip::scoped_lock<ip::interprocess_mutex> lock(shm_dev_->m_mutex);
-    shm_ch_->set_req_read_index(lock, false);
-    return shm_ch_->read_index(lock);
-  };
-
-  void set_write_index(DualIndex new_write_index) override {
-    TimedDualIndex write_index = {new_write_index,
-                                  boost::posix_time::pos_infin};
-    ip::scoped_lock<ip::interprocess_mutex> lock(shm_dev_->m_mutex);
-    shm_ch_->set_req_write_index(lock, false);
-    shm_ch_->set_write_index(lock, write_index);
-  };
-
-  void set_eof(bool eof) override {
-    ip::scoped_lock<ip::interprocess_mutex> lock(shm_dev_->m_mutex);
-    shm_ch_->set_eof(lock, eof);
-  };
-
-  RingBufferView<T_DATA>& data_buffer() override { return *data_buffer_view_; }
-
-  RingBufferView<T_DESC>& desc_buffer() override { return *desc_buffer_view_; }
-
-  DualIndex get_occupied_size() {
-    ip::scoped_lock<ip::interprocess_mutex> lock(shm_dev_->m_mutex);
-    DualIndex read_index = shm_ch_->read_index(lock);
-    DualIndex write_index = shm_ch_->write_index(lock).index;
-    return write_index - read_index;
-  }
+  DualIndex get_occupied_size();
 
   bool empty() { return get_occupied_size() == DualIndex({0, 0}); }
+
+  RingBufferView<T_DATA>& data_buffer() override { return *data_buffer_view_; }
+  RingBufferView<T_DESC>& desc_buffer() override { return *desc_buffer_view_; }
 
 private:
   shm_device* shm_dev_;
